requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
Django==2.1.7
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.3.1
python-google-places==1.4.1
pytz==2018.9
requests==2.21.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.24.1
wrapt==1.11.1
